requirements.txt
fastapi
uvicorn
python-multipart
sqlite3
passlib[bcrypt]
python-jose
pydantic
requests
python-dotenv
groq
